Manually patch FreeRTOS for SystemView integration
SEGGER provides a special patch for FreeRTOS internals in order
to instrument its context switching with special calls that signal
to SystemView when and if a task changes its state while running.

This patch was applied manually as the original patch file expects
a full checkout of FreeRTOS which we don't have here.